Support generation of native installer binaries
Use `./gradlew run` to build and run Bitsquare directly.

Use `./gradlew shadowJar` to generate a portable, executable jar.

Use `./gradlew packageNative` to generate an installer binary specific
to the platform you are running on.

In the case of Windows, it is necessary to copy the two .dll files in
the `package/` directory to the directory where Bitsquare.exe is
ultimately installed.

Resolves #243