Refactor SignatureService
SignatureService methods are now generic with respect to Bitsquare's
domain, i.e. there is no longer any awareness of "contracts",
"registration data", or the like. In fact, because SignatureService
holds no state whatsoever, it would be reasonable to refactor the class
into a set of static utility methods. However, this change leaves
SignatureService as a DI-managed POJO for the time being.

Major changes:

 - Rename #signContract => #signMessage
 - Rename #getEmbeddedAccountRegistrationData => #digestMessageWithSignature