Pin maven-enforcer-plugin version to 1.4.1 (#897)
A new version (v3.0.0-M1) of the enforcer plugin was released on July
27th. If for whatever reason that new version ends up in the local ~/.m2
cache, it results in broken builds because the new version's checksum
gets validated instead of the older v1.4.1 checksum being validated.
This is because the version provided to check was set simply to
'RELEASE', i.e. an alias indicating "whatever the latest release happens
to be". This commit fixes #897 and avoids these potential build failures
by pinning the version explicitly to v1.4.1. This forces the correct
checksum to get validated, in turn causing the build to pass as
expected.