Reformat Java sources with Editorconfig settings
This change is the result of a global reformatting in IDEA now that
`.editorconfig` settings are in place. Note that Editorconfig's
`continuation_indent_size` property [1] was also added and set to 8
here, as this reflects the dominant style already present in the
codebase. i.e. in created a much smaller diff to set the continuation
indent to 8 that would have been created leaving this property out and
letting it default to 4.

[1]: https://github.com/editorconfig/editorconfig/wiki/EditorConfig-Properties