Replace Maven with Gradle in docs and scripts
This commit takes a pass through all documentation and scripts,
replacing any remaining `mvn` commands with their `gradle` equivalents,
replacing `target/*` paths with their `build/*` equivalents, and so on.

Note that the instruction to install Maven has been removed entirely
from documentation and scripts, as opposed to instructing users to
install Gradle, because with the Gradle wrapper (the `gradlew` script in
the root of this repository), it is unnecessary to install Gradle at
all. Users may still do this with, e.g. `brew install gradle` if they
like, but otherwise can get everything they need done with `./gradlew`
commands.

This commit also replaces lowercase 'bisq' with capitalized 'Bisq' where
appropriate in documentation.

Like the previous commit, this is a best-effort approach, and not all
scripts are perfectly up-to-date and tested. Indeed, many of the docs
and scripts we have now are essentially legacy documents and should
probably be scrapped or reworked completely.