Fix bug with not showing popup when database files got corrupted.
The recent changes in the startup broke the sequence if the setup for the CorruptedDataBaseFilesHandler.
We read the data base files before the UI is created and the UI was set as listener to errors. So we missed the errors and did not display a popup.

Fix: Use as CorruptedDataBaseFilesHandler as guice instance to store corrupted files at startup when UI is still not ready. Inject that instance in MainViewModel and shop popup in case there are corrupted files in the list.