Use BlockHash and data for append-only store
We add the blockHash of the 10th block in the break before the relevant phase as part of the hash in the appen-only data.
Together with the date tolerance we get control over the time when the data has been added.
This protects agains various attacks (data withhold and publish late attack).

- Implement DateTolerantPayload at BlindVoteAppendOnlyPayload and ProposalAppendOnlyPayload
- Implement ExpirablePayload at ProposalPayload and BlindVotePayload
- Add blockHash and data at BlindVoteAppendOnlyPayload and ProposalAppendOnlyPayload
- Add ClosedBallotsView