Refactor TxOutput and Block classes
- Add BaseTxOutput and BaseBlock class containing only immutable
blockchain-only data which is common in sub classes
- RawTxOutput extends BaseTxOutput, RawBlock extends BaseBlock
- TempTxOutput extends BaseTxOutput and has mutable data which can be
altered during tx parsing. TempTxOutput is only used during parsing
and gets converted to the immutable TxOutput when tx parsing is
completed.
- TxOutput extends BaseTx and is immutable and used to be stored in
the Block.
- Block extends BaseBlock and is persisted with the BsqState. The tx
list is not a immutable list bu the transactions in there are immutable.
Transactions gets added during parsing to the block which was gets added
to the BsqState when parsing starts.

- Change block time from seconds to ms