Add equals and hashCode methods to classes which use enums as fields
Enums must not be used directly for hashCode or equals as it delivers
the Object.hashCode (internal address)!
The equals and hashCode methods cannot be overwritten in Enums.
It was only applied to classed which are used as value objects and
potentially are used in Sets or other context which involve usage of
equals or hashCode